move update-server-ts to nested folder
